refactor: heroをcomponent化
