+++ src/components/hero/index.ts
@@ -1,0 +1,1 @@
+export { Hero } from './hero'
