+++ src/components/index.ts
@@ -1,3 +1,4 @@
-export { BgBody } from './background'
-export { MineSweeper } from './mine-sweeper'
 export { AccessCount } from './access-count'
+export { BgBody } from './background'
+export { Hero } from './hero'
+export { MineSweeper } from './mine-sweeper'
